// src/permission/enums/permissions.enum.ts
export enum Permissions {
  //* Seller management permissions
  MANAGE_SELLER = 'MANAGE-SELLER',

  //* Customer permissions
  MANAGE_CUSTOMER = 'MANAGE-CUSTOMER',

  //* Transporter permissions
  MANAGE_TRANSPORTER = 'MANAGE-TRANSPORTER',

  //* Product permissions
  MANAGE_PRODUCT = 'MANAGE-PRODUCT',

  //* Platform Revenue permissions
  VIEW_PLATFORM_REVENUE = 'VIEW-PLATFORM-REVENUE',

  //* Reports permissions
  MANAGE_REPORT = 'MANAGE-REPORT',

  //* Payouts permissions
  MANAGE_PAYOUT = 'MANAGE-PAYOUT',

  //* Content management permissions
  MANAGE_CONTENT = 'MANAGE-CONTENT',

  //* Content management permissions
  MANAGE_BRAND = 'MANAGE-BRAND',

  //* ADS management permissions
  MANAGE_AD = 'MANAGE-AD',

  //* ORDERS management permissions
  MANAGE_ORDER = 'MANAGE-ORDER',
}